Quick fix of col.lines
the line col = col.lines[i] (and lty=lty[i],
                     lwd=lwd[i]) were not functioning as ‘i’ is
undefined in the ‘apply’ function. We can pass the index of the line in
question from todraw by adding it as a column. Then it is a simple
matter of defining i as rowdata[5], and the function works as expected.